requirements.txt
asgiref==3.7.2
Brotli==1.1.0
distlib==0.3.7
dj-database-url==2.1.0
Django==3.2.22
django-cors-headers==4.1.0
djangorestframework==3.14.0
filelock==3.12.2
gunicorn==21.2.0
packaging==23.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.5
whitenoise==6.5.0

# pip install asgiref==3.7.2 Brotli==1.1.0 distlib==0.3.7 dj-database-url==2.1.0 Django==3.2.22 django-cors-headers==4.1.0 djangorestframework==3.14.0 filelock==3.12.2 gunicorn==21.2.0 packaging==23.2 platformdirs==3.11.0 psycopg2-binary==2.9.9 pytz==2023.3.post1 sqlparse==0.4.4 typing_extensions==4.7.1 tzdata==2023.3 whitenoise==6.5.0
